test: dyn env, move impl to reducers
