+++ modules/tasks/exec.ts
@@ -91,7 +91,7 @@ export async function execTask(
   args: string[],
   // taskEnv: TaskEnvX,
   // installGraph: InstallGraph,
-): Promise<unknown | undefined> {
+): Promise<Record<string, unknown>> {
   let workSet = new Set([targetKey]);
   {
     const stack = [targetKey];
@@ -109,6 +109,8 @@ export async function execTask(
   if (pendingTasks.length == 0) {
     throw new Error("something went wrong, task graph starting set is empty");
   }
+
+  const output = {} as Record<string, unknown>;
   while (pendingTasks.length > 0) {
     const taskKey = pendingTasks.pop()!;
     const taskDef = tasksConfig.tasks[taskKey];
@@ -153,7 +155,7 @@ export async function execTask(
       ),
     };
 
-    let output: unknown | undefined;
+    let taskOutput: unknown | undefined;
     try {
       if (taskDef.ty == "denoFile@v1") {
         if (!gcx.ghjkfilePath) {
@@ -162,7 +164,7 @@ export async function execTask(
           );
         }
         const workingDir = gcx.ghjkfilePath.parentOrThrow();
-        output = await execTaskDeno(
+        taskOutput = await execTaskDeno(
           gcx.ghjkfilePath.toFileUrl().toString(),
           {
             key: taskDef.key,
@@ -217,9 +219,12 @@ export async function execTask(
     }
     pendingTasks.push(...readyTasks);
 
-    return output;
+    Object.assign(output, {[taskDef.key]: taskOutput});
   }
+
   if (workSet.size > 0) {
     throw new Error("something went wrong, task graph work set is not empty");
   }
+
+  return output;
 }
